Add aliens
- This is the basis for the game, using `GameComponents` as game
subsytems.
- The first thing I've done is create a simple system for moving aliens.
  - As of this commit, they move really fast; this was to test the
  downward/reversing movement.

Next steps:

- The interesting thing is going to be figuring out communication from
the player's bullets to the aliens, so they know when they are hit. I
may use some sort of entity system to signal the creation/destruction of
the game entities. It wouldn't be a real ECS, but it might be the type
of thing that could turn into one eventually.