Convert to MonoGame.Extended ECS
In my research, I discovered there's a very nice library called
MonoGame.Extended. This has a lot of nice extra features, but one of the
most compelling is the Entity Component System.

This allowed me to create a very tightly designed set of subsystems to
do what I was doing before, but with a better separation of concerns. I
created three systems:

- SpriteRenderSystem - This draws all of the sprites in a single batch
to the screen.
- PlayerSystem - Handles player controls
- EnemySystem - Handles motion of the enemy aliens

The rest of the game class is extremely simple, only bootstrapping the
systems, calling world update and draw methods, and loading the sprites
and creating the entities on start.