Add bullet spawn/movement
In this update I added a simple bullet spawn command in PlayerSystem,
and then a simple bullet movement system that moves the bullet to the
top of the screen then despawns it.

The other thing I added was an entity factory, based on the one from a
MonoGame.Extended demo. This was so that I could easily spawn things
from other systems.